theming tests, refactor ItApp as a lib
